Debugging Streamline, Backdrop Added, Cleaning
Debugging streamlined to be less obtrusive. The current scene is a
'snowfall' scene test. Particles will now stick to the floor (and walls
if there is not gravity). Code was cleaned slightly in places. A
backdrop was added to the floor which makes it easy to see what is going
on.